refactor: Coin 클래스 domain 패키지 이동
